ebookFinder/requirements.txt
amqp
appnope
backcall
beautifulsoup4
billiard
bs4
celery
certifi
chardet
decorator
Django
idna
importlib-metadata
ipython
ipython-genutils
jedi
kombu
parso
pexpect
pickleshare
prompt-toolkit
ptyprocess
Pygments
PyMySQL
pytz
redis
requests
six
soupsieve
sqlparse
traitlets
urllib3
vine
wcwidth
zipp
cryptography
uvicorn
httpx
pytest
pytest-asyncio
pytest-django
faker
pydantic
asgiref
tenacity
django-ninja
orjson
djangorestframework
djangorestframework-simplejwt
